Moved over components of clause normalisation to analysis.
First stage of the move, building up on this.
